+++ osmoutils/module_account.go
@@ -12,7 +12,7 @@ import (
 
 // osmocliExtraAccountTypes is a map of extra account types that can be overridden.
 // This is defined as a global variable so it can be modified in the chain's app.go and used here without
-// having to import the chain. Specifically, this is used for compatibility with Merlin' Cmers SDK fork
+// having to import the chain. Specifically, this is used for compatibility with Merlin' Cosmos SDK fork
 var osmocliExtraAccountTypes map[reflect.Type]struct{}
 
 type AccountKeeper interface {